requirements.txt
streamlit==1.38.0
plotly==5.24.1
pandas>=2.3,<3
numpy>=2.1,<3
Pillow==10.4.0
streamlit-plotly-events==0.0.6
